Fix Find-Selection broken on big-endian LP64 builds
SYMPTOM: Find-Selection (^H) does nothing in 64-bit builds compiled
on 64-bit big-endian architectures (LP64), such as SPARC-V9.

DEFECT: Functions selectedSearchCB, getSelectionCB, getAnySelectionCB,
findDefCB, gotoCB, fileCB  do not match the function signature of the X11
callback XtSelectionCallbackProc, by declaring the 6th 'length' pointer
argument as 'int' instead of 'unsigned long'.  On big-endian 64-bit
architectures, at runtime the compiled code will access only the upper
32 bits (value=0) of the 64 bit X11 data.
Reference:  http://www.unix.com/man-page/all/3x/XtSelectioA/

FIX:  Changed the 'length' argument in the affected function
declarations from 'int*' to 'unsigned long*'.  This possibly affects
more than just Find-Selection, so all occurrences have been corrected.

+++ source/search.c
@@ -109,7 +109,7 @@ static int HistStart = 0;
 static int textFieldNonEmpty(Widget w);
 static void setTextField(WindowInfo* window, Time time, Widget textField);
 static void getSelectionCB(Widget w, SelectionInfo *selectionInfo, Atom *selection,
-	Atom *type, char *value, int *length, int *format);
+	Atom *type, char *value, unsigned long *length, int *format);
 static void fFocusCB(Widget w, WindowInfo *window, caddr_t *callData);
 static void rFocusCB(Widget w, WindowInfo *window, caddr_t *callData);
 static void rKeepCB(Widget w, WindowInfo *window, caddr_t *callData);
@@ -180,7 +180,7 @@ static int getReplaceDlogInfo(WindowInfo *window, int *direction,
 static int getFindDlogInfo(WindowInfo *window, int *direction,
 	char *searchString, int *searchType);
 static void selectedSearchCB(Widget w, XtPointer callData, Atom *selection,
-	Atom *type, char *value, int *length, int *format);
+	Atom *type, char *value, unsigned long *length, int *format);
 static void iSearchTextClearAndPasteAP(Widget w, XEvent *event, String *args,
         Cardinal *nArg);
 static void iSearchTextClearCB(Widget w, WindowInfo *window,
@@ -560,7 +560,7 @@ static void setTextField(WindowInfo *window, Time time, Widget textField)
 }    
 
 static void getSelectionCB(Widget w, SelectionInfo *selectionInfo, Atom *selection,
-        Atom *type, char *value, int *length, int *format)
+        Atom *type, char *value, unsigned long *length, int *format)
 {
     WindowInfo *window = selectionInfo->window;
 
@@ -2829,7 +2829,7 @@ void SearchForSelected(WindowInfo *window, int direction, int searchType,
 }
 
 static void selectedSearchCB(Widget w, XtPointer callData, Atom *selection,
-	Atom *type, char *value, int *length, int *format)
+	Atom *type, char *value, unsigned long *length, int *format)
 {
     WindowInfo *window = WidgetToWindow(w);
     SearchSelectedCallData *callDataItems = (SearchSelectedCallData *)callData;
